EES-2127 Replace worker-loader with Webpack 5 web workers
This change also required restructuring of table tool's CSV
util code to avoid the `downloadTableCsv.worker` from having a
module dependency on `xlsx`. This appears to be incompatible with
Webpack 5 for some reason, and causes the frontend build to fail.
