--- src/explore-education-statistics-admin/src/types/worker-loader.d.ts
@@ -1,7 +1,0 @@
-declare module 'worker-loader?name=static/[hash].worker.js!*' {
-  class WebpackWorker extends Worker {
-    public constructor();
-  }
-
-  export default WebpackWorker;
-}
