--- src/explore-education-statistics-common/__mocks__/mock.worker.ts
@@ -1,11 +1,0 @@
-export default class MockWorker {
-  onmessage: () => void;
-
-  constructor() {
-    this.onmessage = () => {};
-  }
-
-  postMessage() {
-    this.onmessage();
-  }
-}
