+++ src/explore-education-statistics-common/package.json
@@ -77,7 +77,6 @@
     "typescript": "^4.4.3",
     "use-immer": "^0.4.0",
     "uuid": "^8.3.1",
-    "worker-loader": "^2.0.0",
     "xlsx": "^0.16.8",
     "yup": "^0.29.3"
   },
