+++ src/explore-education-statistics-common/src/modules/table-tool/components/utils/getCsvData.ts
@@ -6,14 +6,19 @@ import {
   WorkerFullTable,
 } from '@common/modules/table-tool/types/workerFullTable';
 import cartesian from '@common/utils/cartesian';
-import { utils, writeFile } from 'xlsx';
 
 /**
  * Gets the CSV data for download.
- * This is used in the Web Worker so uses a Worker version of the FullTable type (WorkerFullTable). When mapFullTable is refactored to not use classes (EES-2613) we can change this to use the standard FullTable type.
- * IMPORTANT: the filters ids are getters which aren't serialised when passed to the worker so cannot be used here.
+ *
+ * This is used in the Web Worker so uses a Worker version of
+ * the {@see FullTable} type {@see WorkerFullTable}.
+ * When {@see mapFullTable} is refactored to not use classes (EES-2613)
+ * we can change this to use the standard {@see FullTable} type.
+ *
+ * IMPORTANT: the filters ids are getters which aren't serialised when
+ * passed to the worker so cannot be used here.
  */
-export const getCsvData = (fullTable: WorkerFullTable): string[][] => {
+export default function getCsvData(fullTable: WorkerFullTable): string[][] {
   const { subjectMeta, results } = fullTable;
   const { indicators, filters, timePeriodRange, locations } = subjectMeta;
 
@@ -82,14 +87,4 @@ export const getCsvData = (fullTable: WorkerFullTable): string[][] => {
     .filter(row => row.length > 0);
 
   return [columns, ...rows];
-};
-
-export const downloadCsvFile = (csvData: string[][], fileName: string) => {
-  const workBook = utils.book_new();
-  workBook.Sheets.Sheet1 = utils.aoa_to_sheet(csvData);
-  workBook.SheetNames[0] = 'Sheet1';
-
-  writeFile(workBook, `${fileName}.csv`, {
-    type: 'binary',
-  });
-};
+}
