+++ src/explore-education-statistics-common/src/modules/table-tool/components/workers/downloadTableCsv.worker.ts
@@ -1,8 +1,6 @@
-import 'core-js/stable'; // required for IE 11 support
-import 'regenerator-runtime/runtime'; // required for IE 11 support
-import { getCsvData } from '@common/modules/table-tool/components/utils/downloadCsv';
+import getCsvData from '@common/modules/table-tool/components/utils/getCsvData';
 
-/* eslint-disable no-restricted-globals */
+// eslint-disable-next-line no-restricted-globals
 const ctx: Worker = self as never;
 
 /**
@@ -12,4 +10,5 @@ ctx.addEventListener('message', event => {
   const result = getCsvData(event.data.fullTable);
   ctx.postMessage({ csvData: result, fileName: event.data.fileName });
 });
+
 export {};
